src/backend/requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-socketio==5.10.0

# AI/ML
anthropic==0.7.0
cupy-cuda12x==12.3.0
torch==2.1.1
scikit-learn==1.3.2

# Trading & Market Data
alpaca-trade-api==3.0.2
yfinance==0.2.32

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
motor==3.6.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# Web Scraping & News Aggregation
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.0
feedparser==6.0.11
duckduckgo-search==6.3.3
ddgs==1.0.6           # DuckDuckGo Simple client
python-dateutil==2.8.2  # For robust date parsing
pytz==2023.3.post1     # For timezone handling

# Plaid Integration (for Personal Finance Dashboard)
plaid-python==20.0.0

# RAG System
chromadb==0.4.22
sentence-transformers==2.3.1

# Voice (optional providers)
faster-whisper==0.10.1
edge-tts==6.1.12

# Security
pyotp==2.9.0              # TOTP 2FA
cryptography==41.0.7      # Encryption
passlib[bcrypt]==1.7.4    # Password hashing
PyJWT==2.8.0              # JWT tokens

